chore: move canonical popup and styling
