+++ botfront/imports/ui/components/nlu/common/style.import.less
@@ -119,3 +119,23 @@
     }
     line-height: 2.25;
 }
+
+.ui.popup.canonical-popup{
+    max-width: 600px;
+    /* padding is moved from canonical-popup to canonical-popup-content so that
+      the arrow that points to the popup trigger is not removed by overflow: hidden;
+    */
+    padding: 0px; // remove the default padding on the popup
+    .canonical-popup-content {
+      padding: 13.3px 16px 13.3px 16px; // mimic the default padding of the popup
+      display: flex;
+      overflow: hidden;
+    }
+    .icon{
+      margin-right: 0.5em;
+      margin-top: 6px;
+    }
+  }
+  .canonical-popup-trigger{
+    display: inline-block; // match the size of it's children
+  }
